chore(refactor): moved model folder to shared
