+++ web/src/app/core/resolvers/edit-project.resolver.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot, Router } from '@angular/router';
 
 import { AuthenticationService } from '../services/authentication.service';
-import { ProjectModel } from '../../../models/index.model';
+import { ProjectModel } from '../../shared/models/index.model';
 import { ProjectService } from '../services/project.service';
 
 @Injectable({
