+++ web/src/app/core/resolvers/profile.resolver.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Resolve } from '@angular/router';
 
 import { Observable } from 'rxjs';
-import { ProfileModel } from '../../../models/index.model';
+import { ProfileModel } from '../../shared/models/index.model';
 import { first } from 'rxjs/operators';
 import { AuthenticationService } from '../services/authentication.service';
