+++ web/src/app/dialog/list/dialog-list.component.ts
@@ -2,7 +2,7 @@ import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { RepositoryService } from '../../core/services/repository.service';
 import { Subscription } from 'rxjs';
-import { ProjectModel, RepositoriesModel } from '../../../models/index.model';
+import { ProjectModel, RepositoriesModel } from '../../shared/models/index.model';
 
 @Component({
     selector: 'dashboard-dialog-list',
