+++ web/src/app/shared/models/navigation.model.ts
