+++ web/src/app/shared/models/release.model.ts
