+++ web/src/app/shared/models/userStats.model.ts
